fix data type in conversion function
git-svn-id: svn://svn.code.sf.net/p/cxxtools/code/trunk@1193 15ae5fad-cc11-0410-8fac-bce609e504b0

+++ cxxtools/include/cxxtools/convert.h
@@ -261,7 +261,7 @@ inline void convert(double& n, const String& str)
     // not a number
     if(str == L"NAN")
     {
-        n = std::numeric_limits<float>::quiet_NaN();
+        n = std::numeric_limits<double>::quiet_NaN();
         return;
     }
 
@@ -329,7 +329,7 @@ inline void convert(double& n, const std::string& str)
     // not a number
     if(str == "NAN")
     {
-        n = std::numeric_limits<float>::quiet_NaN();
+        n = std::numeric_limits<double>::quiet_NaN();
         return;
     }
